netpulse-ui/requirements.txt
flask
gunicorn
google-adk==1.14.0
toolbox-core
sqlalchemy
pg8000
google-cloud-bigquery
